Add files via upload
Version 4 for the Doomsday Data Dictionary. Updated to reflect changes to vers 2 for the ERD, including the addition of the Resident_Skills linking table. Note additional field changes in the Infected and Patients tables. For the Infected table a Turn_Date (datetime) field has been added. Also, a plethora of new fields have been added to the Patients table to more accurately track the course of the infection (previously we just had a single field with an integer ranking system -- pretty lame, but we knew we'd have to fix it eventually).